refactor: [TG-10] 파일 구조 리팩터링
