+++ frontend/src/components/molecules/auth/IdInputField.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../css/InputField.css'
+import '../../css/InputField.css'
 
 const IdInputField = (props) => {
